(refact) size_t to gint

+++ glib/ml.hpp
@@ -39,7 +39,7 @@ double costLoR(double x, double y) {
 double avgCost(matrix& x, matrix& y, double (*costfunc)(double, double)) {
     gint length = x.getN();                     // # training data
     double sum = 0;
-    for (size_t i = 0; i < length; ++i) {
+    for (gint i = 0; i < length; ++i) {
         sum += costfunc(x[i], y[i]);
     }
     return sum / length / 2;
@@ -52,7 +52,7 @@ double dee(double x) {
 double gradientDesc(matrix& x, matrix& y, matrix& theta, double alpha) {
     gint length = x.getN();
     matrix temp(theta);
-    for (size_t i = 0; i < length; ++i)	{
+    for (gint i = 0; i < length; ++i)	{
         // TODO: complete partial derivative function (dee)
         // If logistic regression, run sigmoid function before calculating cost
         temp[i] = theta[i] - alpha * dee(avgCost(x, y, costLoR));
@@ -80,7 +80,7 @@ public:
     NN(nnConfig cfg) {
         L = cfg.layerSizes.size();
         config = cfg;
-        for (size_t i = 0; i < L; ++i) {
+        for (gint i = 0; i < L; ++i) {
             a.push_back(new matrix(cfg.layerSizes[i]));
         }
         // For consistent layer indexing
@@ -90,7 +90,7 @@ public:
         delta.push_back(nullptr);
         y = new matrix;
 
-        for (size_t i = 1; i < L; ++i) {
+        for (gint i = 1; i < L; ++i) {
             // ith element in a row is the weigth for the ith node in the next layer
             // rows for nodes in the next layer
             // column = weight from node from the prev layer
@@ -118,14 +118,14 @@ public:
         //std::cout << d.min() << ' ' << d.max() << '\n';
 
         for (auto& e : *a[0])               { e = 0; }
-        for (size_t i = 1; i < L; ++i) {
+        for (gint i = 1; i < L; ++i) {
             for (auto& e : *a[i])           { e = d(rng); }
             for (auto& e : *b[i])           { e = d(rng); }
             for (auto& e : *delta[i])       { e = d(rng); }
         }
-        for (size_t i = 1; i < L; ++i) {
+        for (gint i = 1; i < L; ++i) {
             gint length = config.layerSizes[i] * config.layerSizes[i-1];
-            for (size_t j = 0; j < length; ++j) {
+            for (gint j = 0; j < length; ++j) {
                 (*w[i])[j] = 0;
             }
         }
@@ -160,7 +160,7 @@ public:
     // Forward propagation
     void fprop() {
         // For every layer
-        for (size_t i = 1; i < L; ++i) {
+        for (gint i = 1; i < L; ++i) {
             matrix z = *w[i] * *a[i-1] + *b[i];
             *a[i] = sigmoid(z);
         }
@@ -169,14 +169,14 @@ public:
     // Backpropagation
     void bprop() {
         *delta[L-1] = *a[L-1] - *y;
-        for (size_t i = L - 2; i > 0; --i) {
+        for (gint i = L - 2; i > 0; --i) {
             // weights * errors, element multiplied by the derivative of activation values
             *delta[i] = edot(w[i+1]->tr() * *delta[i+1], edot(*a[i], 1.0 - *a[i]));
         }
     }
 
     void partialCostDeltas() {
-        for (size_t i = 0; i < L - 1; ++i) {
+        for (gint i = 0; i < L - 1; ++i) {
             *dee_delta[i] = *dee_delta[i] + *delta[i+1] * a[i]->tr();
         }
     }
@@ -186,7 +186,7 @@ public:
     
     std::string getNumLayers() {
         std::string str;
-        for (size_t i = 0; i < a.size(); ++i) {
+        for (gint i = 0; i < a.size(); ++i) {
             str += std::to_string(a[i]->getLength()) + ' ';
         }
         return str;
@@ -198,15 +198,15 @@ public:
         gint n = config.layerSizes.size();
         gint m = 0;
         gint length;
-        for (size_t i = 1; i < n; ++i) {
+        for (gint i = 1; i < n; ++i) {
             if (config.layerSizes[i] > m) {
                 m = config.layerSizes[i];
             }
         }
 
-        for (size_t j = 0; j < m; ++j) {
+        for (gint j = 0; j < m; ++j) {
             str += "Neuron " + std::to_string(j + 1) + ": ";
-            for (size_t i = 1; i < n; ++i) {
+            for (gint i = 1; i < n; ++i) {
                 length = a[i]->getLength();
                 if (j < length) {
                     str += std::to_string((*a[i])[j]) + "\t";
@@ -221,7 +221,7 @@ public:
     std::string getActivationValues(gint layer) {
         std::string str;
         gint length = a[layer]->getLength();
-        for (size_t i = 0; i < length; ++i)    {
+        for (gint i = 0; i < length; ++i)    {
             str += std::to_string((*a[layer])[i]) + ' ';
         }
         return str;
